refactor: consolidate frontend assets under sites
