+++ docs/agile/tasks/extract_site_modules_from_riatzukiza_github_io_md_md.md
@@ -1,10 +1,11 @@
 ## 🛠️ Task: Extract site modules from riatzukiza.github.io
 
-Port any useful HTML, CSS, or JS from the legacy `riatzukiza.github.io` site into our `/site/` directory.
+Port any useful HTML, CSS, or JS from the legacy `riatzukiza.github.io` site into our `/sites/` directory.
 
 ---
 
 ## 🎯 Goals
+
 - Retain frontend components that describe the project
 - Ensure they build locally under the Promethean repo
 - Phase out obsolete assets
@@ -12,13 +13,15 @@ Port any useful HTML, CSS, or JS from the legacy `riatzukiza.github.io` site int
 ---
 
 ## 📦 Requirements
+
 - [ ] Identify which assets are still relevant
-- [ ] Copy them into `/site/` preserving structure
+- [ ] Copy them into `/sites/` preserving structure
 - [ ] Add README notes on build or deployment steps
 
 ---
 
 ## 📋 Subtasks
+
 - [ ] Review the old repository's site folder
 - [ ] Migrate HTML/CSS/JS files
 - [ ] Adjust paths or links for the new repo layout
@@ -26,18 +29,22 @@ Port any useful HTML, CSS, or JS from the legacy `riatzukiza.github.io` site int
 ---
 
 ## 🔗 Related Epics
+
 #framework-core
 
 ---
 
 ## ⛓️ Blocked By
+
 - Extract docs from riatzukiza.github.io
 
 ## ⛓️ Blocks
+
 Nothing
 
 ---
 
 ## 🔍 Relevant Links
+
 - [kanban](../boards/kanban.md)
-#accepted
+  #accepted
