+++ docs/agile/tasks/migrating_relevant_modules_from_riatzukiza_github_io_to_-site-_and_-docs-_md.md
@@ -5,6 +5,7 @@ Move documentation and site pieces from the legacy `riatzukiza.github.io` reposi
 ---
 
 ## 🎯 Goals
+
 - Preserve useful information from the old repository
 - Keep the project website and docs in a single place
 - Remove outdated or redundant files
@@ -12,32 +13,38 @@ Move documentation and site pieces from the legacy `riatzukiza.github.io` reposi
 ---
 
 ## 📦 Requirements
+
 - [ ] Review the old repo and list files worth keeping
 - [ ] Copy markdown docs into `docs/`
-- [ ] Copy website assets into `site/`
+- [ ] Copy website assets into `sites/`
 - [ ] Update internal links to match the new structure
 
 ---
 
 ## 📋 Subtasks
+
 - [ ] [Extract docs from riatzukiza.github.io](Extract%20docs%20from%20riatzukiza.github.io.md)
 - [ ] [Extract site modules from riatzukiza.github.io](Extract%20site%20modules%20from%20riatzukiza.github.io.md)
 
 ---
 
 ## 🔗 Related Epics
+
 #framework-core
 
 ---
 
 ## ⛓️ Blocked By
+
 Nothing
 
 ## ⛓️ Blocks
+
 - Future website styling tasks
 
 ---
 
 ## 🔍 Relevant Links
+
 - [kanban](../boards/kanban.md)
-#ready
+  #ready
