+++ docs/agile/tasks/migrating_relevant_modules_from_riatzukiza_github_md.md
@@ -1,23 +1,25 @@
-## 🛠️ Task: Migrating relevant modules from `riatzukiza.github.io` to `/site/` and `/docs/`
+## 🛠️ Task: Migrating relevant modules from `riatzukiza.github.io` to `/sites/` and `/docs/`
 
 Generated from [../unique/2025.07.28.18.07.20.md](../unique/2025.07.28.18.07.20.md)
 
 ---
 
 ## 🎯 Goals
+
 - [ ] TODO
 
 ---
 
 ## 📦 Requirements
+
 - [ ] TODO
 
 ---
 
 ## 📋 Subtasks
+
 - [ ] TODO
 
 ---
-
 
 #accepted
