+++ services/ts/llm/src/index.js
@@ -12,7 +12,7 @@ app.use(express.json({ limit: '500mb' }));
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
-app.use(express.static(path.join(__dirname, '../public')));
+app.use(express.static(path.join(__dirname, '../../../../sites/llm-chat')));
 
 let callOllamaFn = async ({ prompt, context, format }, retry = 0) => {
     try {
