+++ sites/README.md
@@ -1,0 +1,11 @@
+# Sites
+
+This directory collects all frontend code for Promethean. Each subfolder contains a standalone HTML/JS frontend served by various services.
+
+Current frontends:
+
+- `smartgpt-dashboard` – OpenAPI-driven dashboard for SmartGPT Bridge.
+- `llm-chat` – Minimal chat interface for the LLM service.
+- `markdown-graph` – Force-directed visualization of markdown link graphs.
+
+Future agents should place any web UI or dashboard code here.
