+++ sites/markdown-graph/README.md
@@ -1,0 +1,3 @@
+# Markdown Link Graph
+
+`graph.html` renders the markdown link graph produced by the `markdown_graph` service. Open the file in a browser while the service is running to explore the connected documents in a force-directed view.
