+++ sites/portfolio/README.md
@@ -6,7 +6,3 @@ The source files have not yet been migrated into the Promethean monorepo.
 For now the folder only contains this README. References to `npm` scripts from the original project have been removed until the site assets and build configuration are restored.
 
 Progress on migrating the site can be tracked in [[MIGRATION_PLAN.md|docs/MIGRATION_PLAN.md]].
-
-## Markdown Link Graph
-
-`graph.html` renders the markdown link graph produced by the `markdown_graph` service. Open the file in a browser while the service is running to explore the connected documents in a force-directed view.
