+++ sites/smartgpt-dashboard/main.js
