+++ sites/smartgpt-dashboard/styles.css
